Cleanup, floats over doubles.

+++ src/collision.h
@@ -5,7 +5,7 @@
 #define WORLD_W 50
 #define WORLD_H 50
 
-#define COLLIDES(x, z) test_collision((x) * ((double) S_W) / D_W, ((z) - Z_OFF) * ((double) S_W) / D_W)
+#define COLLIDES(x, z) test_collision((x) * ((float) S_W) / D_W, ((z) - Z_OFF) * ((float) S_W) / D_W)
 
 bool collision_map[WORLD_H][WORLD_W];
